+++ packages/atomic/src/components/commerce/index.ts
@@ -1,2 +1,4 @@
 // Auto-generated file
 export {AtomicCommerceProductList} from './atomic-commerce-product-list/atomic-commerce-product-list.js';
+export {AtomicCommerceInterface} from './atomic-commerce-interface/atomic-commerce-interface.js';
+export {AtomicCommerceSortDropdown} from './atomic-commerce-sort-dropdown/atomic-commerce-sort-dropdown.js';