+++ packages/atomic/src/components/commerce/lazy-index.ts
@@ -4,6 +4,12 @@ export default {
     await import(
       './atomic-commerce-product-list/atomic-commerce-product-list.js'
     ),
+  'atomic-commerce-interface': async () =>
+    await import('./atomic-commerce-interface/atomic-commerce-interface.js'),
+  'atomic-commerce-sort-dropdown': async () =>
+    await import(
+      './atomic-commerce-sort-dropdown/atomic-commerce-sort-dropdown.js'
+    ),
 } as Record<string, () => Promise<unknown>>;
 
 export type * from './index.js';