+++ packages/atomic/vitest-utils/testing-helpers/fixtures/headless/commerce/search-controller.ts
@@ -1,0 +1,27 @@
+import {Search, SearchState} from '@coveo/headless/commerce';
+
+export const defaultState = {
+  responseId: 'some-id',
+  products: [{}],
+  isLoading: false,
+  error: null,
+};
+export const defaultImplementation = {
+  subscribe: (subscribedFunction: () => void) => {
+    subscribedFunction();
+  },
+  state: defaultState,
+};
+
+export const buildFakeSearch = ({
+  implementation,
+  state,
+}: Partial<{
+  implementation?: Partial<Search>;
+  state?: Partial<SearchState>;
+}>): Search =>
+  ({
+    ...defaultImplementation,
+    ...implementation,
+    ...(state && {state: {...defaultState, ...state}}),
+  }) as Search;